// apps/web/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  SUPERADMIN: 'SUPERADMIN',
  EDITOR: 'EDITOR',
  ORDER_MANAGER: 'ORDER_MANAGER',
  ANALYST: 'ANALYST'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const AddressType = {
  SHIPPING: 'SHIPPING',
  BILLING: 'BILLING'
} as const

export type AddressType = (typeof AddressType)[keyof typeof AddressType]


export const ProductStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED',
  SOLD_OUT: 'SOLD_OUT'
} as const

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus]


export const OrderStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  RETURN_REQUESTED: 'RETURN_REQUESTED',
  RETURNED: 'RETURNED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  AUTHORIZED: 'AUTHORIZED',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED',
  PARTIALLY_REFUNDED: 'PARTIALLY_REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const MovementType = {
  IN: 'IN',
  OUT: 'OUT',
  ADJUSTMENT: 'ADJUSTMENT',
  RETURN: 'RETURN'
} as const

export type MovementType = (typeof MovementType)[keyof typeof MovementType]


export const CouponType = {
  PERCENTAGE: 'PERCENTAGE',
  FIXED: 'FIXED',
  FREE_SHIPPING: 'FREE_SHIPPING'
} as const

export type CouponType = (typeof CouponType)[keyof typeof CouponType]


export const MembershipTier = {
  BLUE: 'BLUE',
  SILVER: 'SILVER',
  GOLD: 'GOLD',
  ELITE: 'ELITE'
} as const

export type MembershipTier = (typeof MembershipTier)[keyof typeof MembershipTier]


export const ReviewVoteType = {
  UP: 'UP',
  DOWN: 'DOWN'
} as const

export type ReviewVoteType = (typeof ReviewVoteType)[keyof typeof ReviewVoteType]


export const BlogStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED'
} as const

export type BlogStatus = (typeof BlogStatus)[keyof typeof BlogStatus]


export const SettingScope = {
  STORE: 'STORE',
  SEO: 'SEO',
  SHIPPING: 'SHIPPING',
  PAYMENT: 'PAYMENT',
  EMAIL: 'EMAIL',
  CMS: 'CMS',
  ANALYTICS: 'ANALYTICS'
} as const

export type SettingScope = (typeof SettingScope)[keyof typeof SettingScope]